Split Config into individual crate Config structs
Allows each crate to define their own configuration options for use by 
their users, thereby keeping configuration easy to maintain (binaries 
can nest a crate's configuration in its own configuration with 
structopt's flatten and they will be reflected in program arguments)

+++ rpc/src/api_definitions.rs
@@ -4,7 +4,7 @@ use serde::{Serialize, Deserialize};
 pub struct Stats {
     pub tail: (u64, String),
     pub target_height: u64,
-    pub difficulty: u64,
+    pub difficulty: u128,
     pub net_type: String,
     pub tx_pool_count: u64
 }
